# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TrabalhoJusto - Plataforma de indicadores ODS 8
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies/{companyId}/trainings:
    get:
      operationId: listTrainings
      tags: [trainings]
      summary: List training programs for a company
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of trainings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Training"
    post:
      operationId: createTraining
      tags: [trainings]
      summary: Add a training program
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrainingBody"
      responses:
        "201":
          description: Created training
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Training"
  /companies/{companyId}/trainings/{id}:
    put:
      operationId: updateTraining
      tags: [trainings]
      summary: Update a training
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrainingBody"
      responses:
        "200":
          description: Updated training
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Training"
    delete:
      operationId: deleteTraining
      tags: [trainings]
      summary: Delete a training
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /curriculos:
    get:
      operationId: listResumes
      tags: [resumes]
      summary: List all resumes
      responses:
        "200":
          description: List of resumes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Resume"
    post:
      operationId: createResume
      tags: [resumes]
      summary: Create a new resume
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResumeBody"
      responses:
        "201":
          description: Created resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
  /curriculos/{id}:
    get:
      operationId: getResume
      tags: [resumes]
      summary: Get a resume by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Resume details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
        "404":
          description: Not found
    put:
      operationId: updateResume
      tags: [resumes]
      summary: Update a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResumeBody"
      responses:
        "200":
          description: Updated resume
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resume"
    delete:
      operationId: deleteResume
      tags: [resumes]
      summary: Delete a resume
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /vagas/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a single job listing
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job listing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
...
  /cursos/{id}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get a single course
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /recomendacoes:
    get:
      operationId: getRecommendations
      tags: [jobs]
      summary: Get recommended jobs and courses based on resume skills
      parameters:
        - name: curriculoId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RecommendationsResponse"
components:
  schemas:
...
    Training:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: integer
        title:
          type: string
        type:
          type: string
          enum: [tecnico, seguranca, lideranca, diversidade, saude, outros]
        hoursTotal:
          type: number
        participantsCount:
          type: integer
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - companyId
        - title
        - type
        - hoursTotal
        - participantsCount
        - createdAt
    CreateTrainingBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [tecnico, seguranca, lideranca, diversidade, saude, outros]
        hoursTotal:
          type: number
        participantsCount:
          type: integer
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - title
        - type
        - hoursTotal
        - participantsCount
...
    Resume:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        city:
          type: string
        state:
          type: string
        linkedinUrl:
          type: string
          nullable: true
        summary:
          type: string
        experiences:
          type: string
          description: JSON array of experience objects
        education:
          type: string
          description: JSON array of education objects
        skills:
          type: string
          description: JSON array of skill objects
        languages:
          type: string
          description: JSON array of language objects
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - phone
        - city
        - state
        - summary
        - experiences
        - education
        - skills
        - languages
        - createdAt
        - updatedAt
    CreateResumeBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        city:
          type: string
        state:
          type: string
        linkedinUrl:
          type: string
          nullable: true
        summary:
          type: string
        experiences:
          type: string
        education:
          type: string
        skills:
          type: string
        languages:
          type: string
      required:
        - name
        - email
        - phone
        - city
        - state
        - summary
        - experiences
        - education
        - skills
        - languages
...
    Job:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        company:
          type: string
        location:
          type: string
        type:
          type: string
        area:
          type: string
        salaryMin:
          type: integer
          nullable: true
        salaryMax:
          type: integer
          nullable: true
        description:
          type: string
        requirements:
          type: string
        skills:
          type: string
        benefits:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - company
        - location
        - type
        - area
        - description
        - requirements
        - skills
        - benefits
        - isActive
        - createdAt
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        provider:
          type: string
        area:
          type: string
        level:
          type: string
        durationHours:
          type: integer
          nullable: true
        description:
          type: string
        skills:
          type: string
        url:
          type: string
          nullable: true
        isFree:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - provider
        - area
        - level
        - description
        - skills
        - isFree
        - createdAt
    RecommendationsResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        courses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
      required:
        - jobs
        - courses