# lib/api-zod/src/generated/types/recommendationsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TrabalhoJusto - Plataforma de indicadores ODS 8
 * OpenAPI spec version: 0.1.0
 */
import type { Course } from "./course";
import type { Job } from "./job";

export interface RecommendationsResponse {
  jobs: Job[];
  courses: Course[];
}